requirements.txt
PyQt6==6.5.3
PyQt6-Qt6==6.5.3
google-cloud-storage==2.14.0
pytest==8.0.0
pytest-mock==3.12.0
pyinstaller==6.4.0
